Restored proper MVC architecture. New Create and insert scripts. DB is now normalized.
