+++ src/main/java/ProjectPortal/Controller/ProjectController.java
@@ -181,86 +181,6 @@ public class ProjectController {
         return "redirect:/" + userId +"/portfolio";
     }
 
-    /**
-     * Handles the update operation for a specific task by retrieving the associated user, project, subproject,
-     * and task details, and adding them to the model for rendering the update task page.
-     * @param userId        the ID of the user who owns the portfolio
-     * @param projectId     the ID of the project the task belongs to
-     * @param subprojectId  the ID of the subproject the task belongs to
-     * @param taskId        the ID of the task to be updated
-     * @param model         the Spring {@code Model} object used to pass attributes to the view
-     * @return the name of the view to render the task update form
-     */
-    @GetMapping("/{userId}/portfolio/{projectId}/{subprojectId}/{taskId}/update")
-    public String updateTask(@PathVariable int userId, @PathVariable int projectId, @PathVariable int subprojectId, @PathVariable int taskId, Model model) {
-        User user = userService.readUserById(userId);
-        Project project = projectService.readProject(projectId);
-        Subproject subproject = subprojectService.readSubproject(subprojectId);
-        Optional<Task> task = taskService.getTaskById(String.valueOf(taskId));
-
-        model.addAttribute("user", user);
-        model.addAttribute("project", project);
-        model.addAttribute("subproject", subproject);
-        model.addAttribute("task", task.get());
-
-        return "update-task";
-    }
-
-    /**
-     * Updates an existing task associated with a specific user, project, and subproject.
-     * @param userId The ID of the user associated with the portfolio.
-     * @param projectId The ID of the project containing the task.
-     * @param subprojectId The ID of the subproject containing the task.
-     * @param taskId The ID of the task to be updated.
-     * @param task The Task object containing the updated details of the task.
-     * @return A redirect URL to the user's portfolio page for the specified project.
-     */
-    @PostMapping("/{userId}/portfolio/{projectId}/{subprojectId}/{taskId}/update")
-    public String updateTask(@PathVariable int userId, @PathVariable int projectId, @PathVariable int subprojectId, @PathVariable int taskId, @ModelAttribute Task task) {
-        task.setSubprojectId(subprojectId);
-        taskService.updateTask(taskId, task);
-        return "redirect:/" + userId + "/portfolio/" + projectId;
-    }
-
-    /**
-     * Handles the update request for a specific subproject within a user's portfolio.
-     * Retrieves user, project, and subproject data based on provided IDs, adds them to the model,
-     * and returns the name of the view for updating the subproject.
-     * @param userId the unique identifier of the user whose subproject is being updated
-     * @param projectId the unique identifier of the project containing the subproject
-     * @param subprojectId the unique identifier of the subproject to be updated
-     * @param model the model object used to supply attributes to the view
-     * @return the name of the view for updating the subproject
-     */
-    @GetMapping("/{userId}/portfolio/{projectId}/{subprojectId}/update")
-    public String updateSubproject(@PathVariable int userId, @PathVariable int projectId, @PathVariable int subprojectId, Model model) {
-        User user = userService.readUserById(userId);
-        Project project = projectService.readProject(projectId);
-        Subproject subproject = subprojectService.readSubproject(subprojectId);
-
-        model.addAttribute("user", user);
-        model.addAttribute("project", project);
-        model.addAttribute("subproject", subproject);
-
-        return "update-subproject";
-    }
-
-    /**
-     * Updates the subproject associated with the given user, project, and subproject IDs
-     * and redirects to the updated project's portfolio view.
-     * @param userId the ID of the user owning the portfolio
-     * @param projectId the ID of the project to which the subproject belongs
-     * @param subprojectId the ID of the subproject to be updated
-     * @param subproject the Subproject object containing the updated subproject details
-     * @return a String that redirects to the portfolio page of the updated project
-     */
-    @PostMapping("/{userId}/portfolio/{projectId}/{subprojectId}/update")
-    public String updateSubproject(@PathVariable int userId, @PathVariable int projectId, @PathVariable int subprojectId, @ModelAttribute Subproject subproject) {
-        subproject.setProjectId(projectId);
-        subprojectService.updateSubproject(subprojectId, subproject);
-        return "redirect:/" + userId + "/portfolio/" + projectId;
-    }
-
 
     /**
      * Deletes a project for a specified user and redirects to the user's portfolio view.
