+++ src/main/resources/create_projectplanner.sql
@@ -30,8 +30,6 @@ CREATE TABLE projects (
                           project_name varchar(255),
                           start_date date NOT NULL,
                           end_date date,
-                          total_estimated_cost double,
-                          total_assigned_employees int,
                           is_complete boolean,
                           project_description text,
                           PRIMARY KEY (project_id),
@@ -45,12 +43,8 @@ CREATE TABLE subprojects (
                              subproject_name varchar(255),
                              start_date date NOT NULL,
                              end_date date,
-                             total_estimated_cost double,
-                             total_assigned_employees int NOT NULL DEFAULT 0,
-                             total_actual_cost double,
                              is_complete boolean,
                              subproject_description text,
-                             hours_allocated int,
                              priority varchar(10),
                              PRIMARY KEY (subproject_id),
                              FOREIGN KEY (project_id) REFERENCES projects (project_id)
